Split ad-init tests into one file per area of functionality
